pyproject.toml
[project]
name = "haiku-mcp"
version = "0.1.0"
description = "MCP Server for generating haikus using AI"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
]

[project.scripts]
haiku-mcp = "haiku_mcp.server:run"
